+++ src/kikuchipy/_constants.py
@@ -27,6 +27,7 @@ deps_for_version_check = [
     "psygnal",
     "pyvista",
     "pyebsdindex",
+    "scikit-optimize",
 ]
 dependency_version: dict[str, Version | None] = {}
 for dep in deps_for_version_check: